Put defaults in IMixinConfigPlugin instead of creating a new interface
